few simple optimizations reduced 5 card GOPS from 131 to 51 sec, less than I thought tbh.
Probably need to store past results so it doesn't repeat, as well as maybe lookup tables (similar to chess).

For example, I'd like to have a lookup table for 2 card GOPS for any card values, based on their relationships (ex. if A is smallest and Z is biggest,
cardA = [A,B], cardB = [B,C],
pointDiff = B prizes = [A,D].
Idk tho, there's already like a lot of combos for just 2 cards.